Autorisations

+++ src/API/views.py
@@ -21,26 +21,46 @@ class RegisterView(generics.CreateAPIView):
 @api_view(["POST","GET","PUT","DELETE"])
 def project_get_post(request,project_id=None):
     project = Projects.objects.filter(id=project_id).first()
+
+    if project_id is None and request.method == 'GET':
+        projects = Projects.objects.all()
+        serializer = ProjectSerializer(projects, many=True)
+        return Response(serializer.data)
+    elif project_id is None and request.method == 'POST':
+        data = request.data
+        serializer = ProjectSerializer(data=data, context={'request': request})
+        if serializer.is_valid():
+            serializer.save()
+            return JsonResponse(serializer.data, status=204)
+        return JsonResponse(serializer.errors, status=400)
+    elif project_id is None:
+        return JsonResponse({"erreur": "Requête incorrecte"}, status=401)
+
     if project:
         if not has_project_permission(request,project_id):
             return JsonResponse({"erreur": "Vous n'avez pas l'autorisation d'accès"}, status=401)
         if request.method == 'GET':
-            if project_id != None:
-                if project:
-                    serializer = ProjectSerializer(project)
-                    return Response(serializer.data)
-            else:
-                projects = Projects.objects.all()
-                serializer = ProjectSerializer(projects,many=True)
+            if project:
+                serializer = ProjectSerializer(project)
                 return Response(serializer.data)
 
-        elif request.method == 'POST':
-            data = request.data
-            serializer = ProjectSerializer(data=data,context={'request': request})
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data, status=204)
-            return JsonResponse(serializer.errors, status=400)
+        elif request.method == 'PUT':
+            if is_object_author(request,project):
+                data = request.data
+                serializer = ProjectSerializer(project, data=data, partial=True)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data)
+                return JsonResponse(serializer.errors, status=400)
+            else:
+                return JsonResponse({"erreur": "Vous n'avez pas l'autorisation de mise à jour"}, status=401)
+
+        elif request.method == 'DELETE':
+            if is_object_author(request,project):
+                project.delete()
+                return JsonResponse({"détails": "Projet {} supprimé".format(project_id)},status=200)
+            else:
+                return JsonResponse({"erreur": "Vous n'avez pas l'autorisation de suppression"}, status=401)
     else:
         return JsonResponse({"erreur":"Projet Inexistant"})
 
@@ -58,6 +78,15 @@ def has_project_permission(request,project_id):
     if contrib_or_user :
         return True
     return False
+
+def is_object_author(request,object):
+    if object.author_user_id is None:
+        return False
+    else:
+        if request.user.id == object.author_user_id:
+            return True
+        return False
+
 
 @csrf_exempt
 @api_view(["POST","GET","PUT","DELETE"])
@@ -147,21 +176,27 @@ def update_issue(request,project_id,issue_id):
 
     if request.method == 'DELETE':
         issue = Issues.objects.filter(project_id=project_id,id=issue_id)
-        if issue:
-            issue.delete()
+        if is_object_author(request, issue):
+            if issue:
+                issue.delete()
+            else:
+                return JsonResponse({"erreur":"Aucun Problème à supprimer"})
+            return JsonResponse({"détail":"Le projet {} a été supprimé".format(issue_id)})
         else:
-            return JsonResponse({"erreur":"Aucun Problème à supprimer"})
-        return JsonResponse({"détail":"Le projet {} a été supprimé".format(issue_id)})
+            return JsonResponse({"erreur": "Vous n'avez pas l'autorisation de suppression"}, status=401)
 
     elif request.method == 'PUT':
         issue = Issues.objects.filter(project_id=project_id, id=issue_id).first()
         if issue:
-            data = request.data
-            serializer = IssueSerializer(issue,data=data, partial=True)
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data)
-            return JsonResponse(serializer.errors, status=400)
+            if is_object_author(request, issue):
+                data = request.data
+                serializer = IssueSerializer(issue,data=data, partial=True)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data)
+                return JsonResponse(serializer.errors, status=400)
+            else:
+                return JsonResponse({"erreur": "Vous n'avez pas l'autorisation de mise à jour"},status=401)
         else:
             return JsonResponse({"erreur": "Aucun Problème correspondant"})
 
@@ -202,6 +237,17 @@ def put_delete_comments(request,project_id,issue_id,comment_id):
     if not has_project_permission(request,project_id):
         return JsonResponse({"erreur": "Vous n'avez pas l'autorisation d'accès"}, status=401)
 
+    if request.method == 'GET':
+        comment = Comments.objects.filter(id=comment_id)
+        if comment:
+            serializer = CommentSerializer(comment, many=True, context={'request': request, 'issue_id': issue_id})
+            return Response(serializer.data)
+        else:
+            return JsonResponse({"erreur": "Aucun Commentaire avec cet ID"})
+
+    if not is_object_author(request,Comments.objects.filter(id=comment_id).first()):
+        return JsonResponse({"erreur": "Vous n'avez pas l'autorisation de suppression ou mise à jour"}, status=401)
+
     if existing_issue(project_id,issue_id):
         if request.method == 'DELETE':
             comment = Comments.objects.filter(id=comment_id)
@@ -222,14 +268,6 @@ def put_delete_comments(request,project_id,issue_id,comment_id):
                 return JsonResponse(serializer.errors, status=400)
             else:
                 return JsonResponse({"erreur": "Aucun Commentaire correspondant"})
-
-        elif request.method == 'GET':
-            comment = Comments.objects.filter(id=comment_id)
-            if comment:
-                serializer = CommentSerializer(comment,many=True, context={'request': request, 'issue_id': issue_id})
-                return Response(serializer.data)
-            else:
-                return JsonResponse({"erreur": "Aucun Commentaire avec cet ID"})
     else:
         return JsonResponse({"erreur": 'Projet ou problème inexistant'})
